# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "YAOpenEvidence API",
    "description": "YAOpenEvidence 医学文献证据问答 API。\n\n- 问答任务异步执行：`POST /v1/answers` 返回 202，进度走 `GET /v1/jobs/{id}/events`（SSE）\n- 鉴权：管理员创建账号，登录后使用 `Authorization: Bearer <access_token>`（包括 SSE）\n- 错误：RFC 9457 Problem Details（`application/problem+json`），按 `code` 分支\n",
    "version": "0.1.0"
  },
  "paths": {
...
    "/v1/auth/login": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "使用账号密码登录",
        "operationId": "login_v1_auth_login_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoginResponse"
                }
              }
            }
          },
          "default": {
            "description": "Problem Details (RFC 9457)",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "422": {
            "description": "Request validation failed",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      }
    },
    "/v1/auth/logout": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "注销本次会话",
        "operationId": "logout_v1_auth_logout_post",
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "default": {
            "description": "Problem Details (RFC 9457)",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "422": {
            "description": "Request validation failed",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/v1/auth/me": {
      "get": {
        "tags": [
          "auth"
        ],
        "summary": "读取当前账号",
        "operationId": "me_v1_auth_me_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserRead"
                }
              }
            }
          },
          "default": {
            "description": "Problem Details (RFC 9457)",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "422": {
            "description": "Request validation failed",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/v1/auth/password": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "修改密码并注销全部会话",
        "operationId": "change_password_v1_auth_password_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PasswordChangeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "default": {
            "description": "Problem Details (RFC 9457)",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "422": {
            "description": "Request validation failed",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
...
    "/v1/jobs": {
      "get": {
        "tags": [
          "jobs"
        ],
        "summary": "任务列表",
        "operationId": "list_jobs_v1_jobs_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "kind",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "enum": [
                    "ask",
                    "kb_reindex"
                  ],
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Kind"
            }
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "enum": [
                    "queued",
                    "running",
                    "succeeded",
                    "failed",
                    "cancelled"
                  ],
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 20,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page_Job_"
                }
              }
            }
          },
          "default": {
            "description": "Problem Details (RFC 9457)",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "422": {
            "description": "Request validation failed",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      }
    },
    "/v1/jobs/{job_id}": {
      "get": {
        "tags": [
          "jobs"
        ],
        "summary": "任务详情",
        "operationId": "get_job_v1_jobs__job_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Job Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Job"
                }
              }
            }
          },
          "default": {
            "description": "Problem Details (RFC 9457)",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "422": {
            "description": "Request validation failed",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      }
    },
    "/v1/jobs/{job_id}/cancel": {
      "post": {
        "tags": [
          "jobs"
        ],
        "summary": "请求取消任务",
        "operationId": "cancel_job_v1_jobs__job_id__cancel_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Job Id"
            }
          }
        ],
        "responses": {
          "202": {
            "description": "Successful Response",
            "headers": {
              "Location": {
                "schema": {
                  "type": "string"
                },
                "description": "任务状态地址"
              }
            }
          },
          "default": {
            "description": "Problem Details (RFC 9457)",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "422": {
            "description": "Request validation failed",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      }
    },
    "/v1/jobs/{job_id}/events": {
      "get": {
        "tags": [
          "jobs"
        ],
        "summary": "任务进度流（SSE）",
        "operationId": "job_events_v1_jobs__job_id__events_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Job Id"
            }
          },
          {
            "name": "Last-Event-ID",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "maxLength": 41,
              "pattern": "^(0|[1-9][0-9]{0,19})-(0|[1-9][0-9]{0,19})$",
              "description": "上次已处理的 Stream ID：两个无前导零的无符号 64 位整数，以连字符分隔。",
              "default": "0-0",
              "title": "Last-Event-Id"
            },
            "description": "上次已处理的 Stream ID：两个无前导零的无符号 64 位整数，以连字符分隔。"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "text/event-stream": {
                "schema": {
                  "type": "string",
                  "description": "SSE 帧；data 为 JSON，模型见 x-sse-events。"
                },
                "x-sse-events": {
                  "stage": {
                    "$ref": "#/components/schemas/StageEventData"
                  },
                  "progress": {
                    "$ref": "#/components/schemas/ProgressEventData"
                  },
                  "log": {
                    "$ref": "#/components/schemas/LogEventData"
                  },
                  "succeeded": {
                    "$ref": "#/components/schemas/SucceededEventData"
                  },
                  "failed": {
                    "$ref": "#/components/schemas/FailedEventData"
                  },
                  "cancelled": {
                    "$ref": "#/components/schemas/CancelledEventData"
                  }
                }
              }
            }
          },
          "default": {
            "description": "Problem Details (RFC 9457)",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "422": {
            "description": "Request validation failed",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Job": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "kind": {
            "type": "string",
            "enum": [
              "ask",
              "kb_reindex"
            ],
            "title": "Kind"
          },
          "status": {
            "type": "string",
            "enum": [
              "queued",
              "running",
              "succeeded",
              "failed",
              "cancelled"
            ],
            "title": "Status"
          },
          "user_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "User Id"
          },
          "params": {
            "additionalProperties": true,
            "type": "object",
            "title": "Params",
            "default": {}
          },
          "progress": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/JobProgress"
              },
              {
                "type": "null"
              }
            ]
          },
          "error": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/JobError"
              },
              {
                "type": "null"
              }
            ]
          },
          "result": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Result"
          },
          "created_at": {
            "type": "string",
            "title": "Created At"
          },
          "started_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Started At"
          },
          "finished_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Finished At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "kind",
          "status",
          "created_at"
        ],
        "title": "Job"
      },
      "JobError": {
        "properties": {
          "code": {
            "type": "string",
            "title": "Code"
          },
          "message": {
            "type": "string",
            "title": "Message"
          }
        },
        "type": "object",
        "required": [
          "code",
          "message"
        ],
        "title": "JobError"
      },
      "JobProgress": {
        "properties": {
          "stage": {
            "type": "string",
            "title": "Stage"
          },
          "current": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Current"
          },
          "total": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Total"
          }
        },
        "type": "object",
        "required": [
          "stage"
        ],
        "title": "JobProgress"
      },
...
      "LoginRequest": {
        "properties": {
          "username": {
            "type": "string",
            "maxLength": 64,
            "minLength": 3,
            "pattern": "^[A-Za-z0-9][A-Za-z0-9_.-]{2,63}$",
            "title": "Username"
          },
          "password": {
            "type": "string",
            "maxLength": 128,
            "minLength": 12,
            "format": "password",
            "title": "Password",
            "writeOnly": true
          }
        },
        "type": "object",
        "required": [
          "username",
          "password"
        ],
        "title": "LoginRequest"
      },
      "LoginResponse": {
        "properties": {
          "access_token": {
            "type": "string",
            "title": "Access Token"
          },
          "token_type": {
            "type": "string",
            "const": "bearer",
            "title": "Token Type",
            "default": "bearer"
          },
          "expires_at": {
            "type": "string",
            "title": "Expires At"
          },
          "user": {
            "$ref": "#/components/schemas/UserRead"
          }
        },
        "type": "object",
        "required": [
          "access_token",
          "expires_at",
          "user"
        ],
        "title": "LoginResponse"
      },
...
      "Page_Job_": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/Job"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "limit": {
            "type": "integer",
            "title": "Limit"
          },
          "offset": {
            "type": "integer",
            "title": "Offset"
          }
        },
        "type": "object",
        "required": [
          "items",
          "total",
          "limit",
          "offset"
        ],
        "title": "Page[Job]"
      },
...
      "PasswordChangeRequest": {
        "properties": {
          "current_password": {
            "type": "string",
            "maxLength": 128,
            "minLength": 12,
            "format": "password",
            "title": "Current Password",
            "writeOnly": true
          },
          "new_password": {
            "type": "string",
            "maxLength": 128,
            "minLength": 12,
            "format": "password",
            "title": "New Password",
            "writeOnly": true
          }
        },
        "type": "object",
        "required": [
          "current_password",
          "new_password"
        ],
        "title": "PasswordChangeRequest"
      },
...
      "UserRead": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "username": {
            "type": "string",
            "title": "Username"
          },
          "role": {
            "type": "string",
            "enum": [
              "user",
              "admin"
            ],
            "title": "Role"
          },
          "is_active": {
            "type": "boolean",
            "title": "Is Active"
          },
          "created_at": {
            "type": "string",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "username",
          "role",
          "is_active",
          "created_at"
        ],
        "title": "UserRead"
      },
...
      "Problem": {
        "description": "RFC 9457 Problem Details。`code` 是稳定枚举，`detail` 仅供人读。",
        "properties": {
          "type": {
            "examples": [
              "urn:yaoe:error:not_found"
            ],
            "title": "Type",
            "type": "string"
          },
          "title": {
            "examples": [
              "Not Found"
            ],
            "title": "Title",
            "type": "string"
          },
          "status": {
            "examples": [
              404
            ],
            "title": "Status",
            "type": "integer"
          },
          "detail": {
            "examples": [
              "answer 'abc' not found"
            ],
            "title": "Detail",
            "type": "string"
          },
          "instance": {
            "examples": [
              "/v1/answers/abc"
            ],
            "title": "Instance",
            "type": "string"
          },
          "code": {
            "examples": [
              "not_found"
            ],
            "title": "Code",
            "type": "string"
          },
          "errors": {
            "anyOf": [
              {
                "items": {
                  "additionalProperties": true,
                  "type": "object"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Errors"
          }
        },
        "required": [
          "type",
          "title",
          "status",
          "detail",
          "instance",
          "code"
        ],
        "title": "Problem",
        "type": "object"
      },
      "AnswerSucceededData": {
        "additionalProperties": false,
        "properties": {
          "answer_id": {
            "title": "Answer Id",
            "type": "string"
          }
        },
        "required": [
          "answer_id"
        ],
        "title": "AnswerSucceededData",
        "type": "object"
      },
      "CancelledEventData": {
        "additionalProperties": false,
        "properties": {},
        "title": "CancelledEventData",
        "type": "object"
      },
      "FailedEventData": {
        "additionalProperties": false,
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          },
          "message": {
            "title": "Message",
            "type": "string"
          }
        },
        "required": [
          "code",
          "message"
        ],
        "title": "FailedEventData",
        "type": "object"
      },
      "LogEventData": {
        "additionalProperties": false,
        "properties": {
          "level": {
            "enum": [
              "info",
              "warning"
            ],
            "title": "Level",
            "type": "string"
          },
          "message": {
            "title": "Message",
            "type": "string"
          }
        },
        "required": [
          "level",
          "message"
        ],
        "title": "LogEventData",
        "type": "object"
      },
      "ProgressEventData": {
        "additionalProperties": false,
        "properties": {
          "stage": {
            "enum": [
              "fulltext",
              "read",
              "kb",
              "reindex"
            ],
            "title": "Stage",
            "type": "string"
          },
          "current": {
            "minimum": 0,
            "title": "Current",
            "type": "integer"
          },
          "total": {
            "minimum": 0,
            "title": "Total",
            "type": "integer"
          },
          "pmid": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Pmid"
          },
          "title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Title"
          }
        },
        "required": [
          "stage",
          "current",
          "total"
        ],
        "title": "ProgressEventData",
        "type": "object"
      },
      "ReindexSucceededData": {
        "additionalProperties": false,
        "properties": {
          "items": {
            "minimum": 0,
            "title": "Items",
            "type": "integer"
          },
          "papers": {
            "minimum": 0,
            "title": "Papers",
            "type": "integer"
          }
        },
        "required": [
          "items",
          "papers"
        ],
        "title": "ReindexSucceededData",
        "type": "object"
      },
      "StageEventData": {
        "additionalProperties": false,
        "properties": {
          "stage": {
            "enum": [
              "queries",
              "search",
              "fulltext",
              "read",
              "kb",
              "synthesize",
              "reindex"
            ],
            "title": "Stage",
            "type": "string"
          },
          "status": {
            "enum": [
              "started",
              "finished"
            ],
            "title": "Status",
            "type": "string"
          },
          "detail": {
            "additionalProperties": true,
            "title": "Detail",
            "type": "object"
          }
        },
        "required": [
          "stage",
          "status"
        ],
        "title": "StageEventData",
        "type": "object"
      },
      "SucceededEventData": {
        "anyOf": [
          {
            "$ref": "#/components/schemas/AnswerSucceededData"
          },
          {
            "$ref": "#/components/schemas/ReindexSucceededData"
          }
        ],
        "title": "SucceededEventData"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "description": "登录返回的可撤销会话令牌",
        "scheme": "bearer"
      }
    }
  }
}